🔨🤖 Move non-indicator helpers out of metadataHelpers
metadataHelpers goes from 17 exports to 13. Three of them had nothing to do
with indicator metadata:

- formatAuthors and formatAuthorsForBibtex format gdoc bylines. They move to
  Util.ts, beside formatInlineList. This takes remeda out of metadataHelpers.
- getPhraseForArchivalDate joins parseArchivalDate, which it calls, in
  archival/archivalDate.ts. The private getDateForCitation went with it as
  formatDateForCitation, since getCitationLong needs it too and
  metadataHelpers already imported from that module.
- splitSourceTextIntoFragments had one consumer, so it now lives in
  IndicatorKeyData.tsx and is no longer a package export. It is exported from
  the module rather than the components index, to keep the package surface
  unchanged.

Sitting next to formatInlineList showed that formatAuthors was that function
plus an Oxford comma, so it is now a call to it with the comma switched on.
The comma applies from three items up, so two-item lists still read "A and B"
and the output is identical for every length. formatInlineList's connector
moves into an options object next to the new oxfordComma flag, and its callers
follow.

No behaviour change anywhere. Tests follow the functions into Util.test.ts and
archivalDate.test.ts.
