+++ packages/@ourworldindata/utils/src/archival/archivalDate.test.ts
@@ -1,5 +1,9 @@
 import { afterEach, beforeEach, describe, expect, it } from "vitest"
-import { formatAsArchivalDate, parseArchivalDate } from "./archivalDate.js"
+import {
+    formatAsArchivalDate,
+    getPhraseForArchivalDate,
+    parseArchivalDate,
+} from "./archivalDate.js"
 import dayjs from "../dayjs.js"
 import timezoneMock from "timezone-mock"
 
@@ -56,5 +60,31 @@ describe(formatAsArchivalDate, () => {
         const date = dayjs("2020-01-21T12:34:56Z")
         const formattedDate = formatAsArchivalDate(date)
         expect(formattedDate).toBe("20200121-123456")
+    })
+})
+
+describe(getPhraseForArchivalDate, () => {
+    it("returns undefined if there is no archival date", () => {
+        expect(getPhraseForArchivalDate(undefined)).toBeUndefined()
+    })
+
+    it("formats an archival timestamp", () => {
+        expect(getPhraseForArchivalDate("20250414-074331")).toEqual(
+            "(archived on April 14, 2025)."
+        )
+    })
+
+    it("formats an ISO date", () => {
+        expect(getPhraseForArchivalDate("2025-04-14")).toEqual(
+            "(archived on April 14, 2025)."
+        )
+    })
+
+    // TODO: return undefined when the date cannot be parsed, rather than
+    // putting "Invalid Date" into a citation.
+    it("renders an invalid date for an unparseable archival date", () => {
+        expect(getPhraseForArchivalDate("nope")).toEqual(
+            "(archived on Invalid Date)."
+        )
     })
 })
